Move Comments feature to vertical-slice folders
Pilot the vertical-slice folder convention with the Comments
feature. Controller, service, interface, DTOs, route constants,
and tests now live under Features/Comments/ in their respective
projects. Other features unchanged.

- ICommentService moved out of Domain.Services into
  Application.Features.Comments — interface is the feature's
  contract, not a domain abstraction.
- CommentRoutes extracted from ApiEndpoints.cs.
- DI registration moved into per-feature AddCommentsFeature()
  extension in CommentsModule.cs.
- Entity (Comment) and DbContext stay in Domain/Infrastructure
  as before.
