+++ src/backend/TB.DanceDance.API.Contracts/Features/Comments/CommentResponse.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace TB.DanceDance.API.Contracts.Responses
+namespace TB.DanceDance.API.Contracts.Features.Comments
 {
     public class CommentResponse
     {
