+++ src/backend/TB.DanceDance.API.Contracts/Features/Comments/Comments.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace TB.DanceDance.API.Contracts.Requests
+namespace TB.DanceDance.API.Contracts.Features.Comments
 {
     public abstract class BaseComment
     {
