+++ src/tests/TB.DanceDance.Tests/Features/Comments/CommentServiceTests.cs
@@ -1,10 +1,9 @@
-using Application.Services;
+using Application.Features.Comments;
 using Domain.Entities;
-using Domain.Services;
 using Infrastructure.Data;
 using TB.DanceDance.Tests.TestsFixture;
 
-namespace TB.DanceDance.Tests.Application;
+namespace TB.DanceDance.Tests.Features.Comments;
 
 public class CommentServiceTests : BaseTestClass
 {
